+++ .prettierignore
@@ -2,7 +2,7 @@
 node_modules
 public
 .github
-pages/global-css.ts
+src/global-css.ts
 
 test/examples
 src/devtools/client/debugger/src/workers/parser/tests/fixtures
